+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/controller/ControllerBlockItem.java
@@ -55,7 +55,7 @@ public class ControllerBlockItem extends AbstractEnergyBlockItem {
     public void appendHoverText(final ItemStack stack, final TooltipContext context, final TooltipDisplay display,
                                 final Consumer<Component> builder, final TooltipFlag flag) {
         super.appendHoverText(stack, context, display, builder, flag);
-        if (!Platform.INSTANCE.getConfig().isRequireEnergy()) {
+        if (!RefinedStorageApi.INSTANCE.isEnergyRequired()) {
             builder.accept(NOT_NEEDED);
         }
     }